Refactored layouts module into its own file
